+++ tests/unit/RunnerTest.php
@@ -19,12 +19,17 @@ class RunnerTest extends UnitTestCase
      * @see setUp for initialization
      */
     private $deploy_copy_cmd;
+    private $deploy_copy_cmd_2;
 
     protected function setUp()
     {
         parent::setUp();
 
-        $this->deploy_copy_cmd = "cd /tmp/pipelines-test-suite/. " .
+        $this->deploy_copy_cmd = "cd /tmp/pipelines/cp/. " .
+            "&& echo 'app' | tar c -h -f - --no-recursion app " .
+            "| docker  cp - '*dry-run*:/.'";
+
+        $this->deploy_copy_cmd_2 = "cd /tmp/pipelines-test-suite/. " .
             "&& tar c -f - . " .
             "| docker  cp - '*dry-run*:/app'";
     }
@@ -149,6 +154,7 @@ class RunnerTest extends UnitTestCase
         $exec
             ->expect('capture', 'docker', 0)
             ->expect('pass', $this->deploy_copy_cmd, 0)
+            ->expect('pass', $this->deploy_copy_cmd_2, 0)
             ->expect('pass', 'docker', 0)
         ;
 
@@ -241,6 +247,7 @@ class RunnerTest extends UnitTestCase
         $exec
             ->expect('capture', 'docker', 0)
             ->expect('pass', $this->deploy_copy_cmd, 0)
+            ->expect('pass', $this->deploy_copy_cmd_2, 0)
             ->expect('pass', 'docker', 0)
             ->expect('capture', 'docker',"./build/foo-package.tgz")
             ->expect('pass', 'docker exec -w /app \'*dry-run*\' tar c -f - build/foo-package.tgz | tar x -f - -C /tmp/pipelines-test-suite', 0)
@@ -275,6 +282,7 @@ class RunnerTest extends UnitTestCase
         $exec
             ->expect('capture', 'docker', 0)
             ->expect('pass', $this->deploy_copy_cmd, 0)
+            ->expect('pass', $this->deploy_copy_cmd_2, 0)
             ->expect('pass', 'docker', 0)
             ->expect('capture', 'docker',"./build/foo-package.tgz")
         ;
@@ -308,6 +316,7 @@ class RunnerTest extends UnitTestCase
         $exec
             ->expect('capture', 'docker', 0)
             ->expect('pass', $this->deploy_copy_cmd, 0)
+            ->expect('pass', $this->deploy_copy_cmd_2, 0)
             ->expect('pass', 'docker', 0)
             ->expect('capture', 'docker',"./build/foo-package.tgz")
             ->expect('pass', 'docker exec -w /app \'*dry-run*\' tar c -f - build/foo-package.tgz | tar x -f - -C /tmp/pipelines-test-suite', 1)